Add final selection page, integrate lobbies with restaurant selections, add voting and scoring (#17)
Updated lobby in redux state to represent 1 lobby instead of an array
Updated actions and added missing actions where server state was out of sync with redux state
When clicking 'start search' on /lobbypage/ will now send a query with filters to yelp api and return a real list of restaurants. Filter happens in 2 steps 1) the filters that yelp uses we can just supply as part of the query, 2) take the finished data and filter it again on the returned results. It seems like there are filters we could remove (e.g. max reviews or min distance)
Vote buttons now add the votes to the lobby. When all participants have voted, then we will calculate the restaurant with the most 'yes' votes. ties will be broken arbitrarily.
Final selection page now appears after all votes are cast. We check in 2s intervals to see if voting is completed.